requirements.txt
altgraph==0.15
Django==2.0.2
future==0.16.0
macholib==1.9
pefile==2017.11.5
Pillow==5.0.0
PyInstaller==3.3.1
pypiwin32==219
PyQt5==5.10
pyqt5-tools==5.9.0.1.2
python-dateutil==2.6.1
pytz==2017.3
sip==4.19.7
six==1.11.0
Unipath==1.1
xlrd==1.1.0
